+++ Solution/AirplaneShooter/DataTables.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Aircraft.h"
+#include "Render\Aircraft.h"
 #include "Ids\TextureId.h"
 
 #include <vector>
