+++ Solution/AirplaneShooter/Render/Projectile.cpp
@@ -1,0 +1,6 @@
+#include "Projectile.h"
+
+using namespace Shooter;
+
+Projectile::Projectile() {}
+Projectile::~Projectile() {}
