+++ Solution/AirplaneShooter/Render/Projectile.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include "Entity.h"
+
+namespace Shooter {
+
+	class Projectile : public Entity {
+	public:
+		Projectile();
+		~Projectile();
+	};
+
+}
