+++ Solution/Game2D/ResourceManager.h
