+++ Solution/Game2D/ResourceManager.inl
